Move unit test files
Mirror the source code file structure in the unit test organization.
